backend/requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
aiohttp==3.9.1
lxml==4.9.3
requests>=2.26.0

# Backend API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.3
python-multipart>=0.0.5

# Database 
sqlalchemy==2.0.25
motor==3.3.2  # (If using MongoDB)
pymongo==4.6.0
supabase==1.2.0  # Downgraded to a stable version
httpx==0.24.1  # Specific version for compatibility

# Machine Learning (Fish Classification & Compatibility Assessment)
numpy>=1.19.0
scipy==1.11.4
pandas>=1.3.0
scikit-learn>=0.24.0
tensorflow==2.15.0
torch==2.1.2
torchvision==0.16.2  # Updated to match PyTorch version
opencv-python==4.9.0.80
matplotlib==3.8.2

# User Authentication & Security (Admin Controls)
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Miscellaneous
python-dotenv==1.0.1
Pillow==10.1.0
email-validator==2.1.0.post1
typer==0.9.0
psycopg2-binary==2.9.9
joblib==1.3.2

